+++ src/main/java/ru/icc/cells/tabbypdf/utils/processing/filter/bi/HorizontalPositionBiHeuristic.java
@@ -4,21 +4,17 @@ import ru.icc.cells.tabbypdf.common.Rectangle;
 import ru.icc.cells.tabbypdf.common.TextBlock;
 import ru.icc.cells.tabbypdf.common.TextChunk;
 
-public class HorizontalPositionBiHeuristic extends BiHeuristic<Rectangle>
-{
+public class HorizontalPositionBiHeuristic extends BiHeuristic<Rectangle> {
 
-    public HorizontalPositionBiHeuristic()
-    {
+    public HorizontalPositionBiHeuristic() {
         super(Orientation.HORIZONTAL);
     }
 
     @Override
-    public boolean test(Rectangle first, Rectangle second)
-    {
-        float lx1, lx2, ty1, ty2, by1, by2;
-        if (first.getClass().equals(TextBlock.class) && second.getClass().equals(TextBlock.class))
-        {
-            TextChunk firstChunk  = ((TextBlock) first).getChunks().get(((TextBlock) first).getChunks().size() - 1);
+    public boolean test(Rectangle first, Rectangle second) {
+        double lx1, lx2, ty1, ty2, by1, by2;
+        if (first.getClass().equals(TextBlock.class) && second.getClass().equals(TextBlock.class)) {
+            TextChunk firstChunk = ((TextBlock) first).getChunks().get(((TextBlock) first).getChunks().size() - 1);
             TextChunk secondChunk = ((TextBlock) second).getChunks().get(0);
             lx1 = firstChunk.getLeft();
             lx2 = secondChunk.getLeft();
@@ -26,9 +22,7 @@ public class HorizontalPositionBiHeuristic extends BiHeuristic<Rectangle>
             ty2 = secondChunk.getTop();
             by1 = firstChunk.getBottom();
             by2 = secondChunk.getBottom();
-        }
-        else
-        {
+        } else {
             lx1 = first.getLeft();
             lx2 = second.getLeft();
             ty1 = first.getTop();
